+++ frontend/components/Web/Forms/FormErrorParser.ts
@@ -7,8 +7,6 @@
  * file that was distributed with this source code.
  */
 
-import {GraphQLError} from "graphql";
-
 const FormErrorParser = () => {
     const parse = (errorMessage: string): string => {
         return errorMessage.includes(':') ? errorMessage.substring(errorMessage.indexOf(':') + 2) : errorMessage;
