+++ frontend/lib/api.js
@@ -8,13 +8,13 @@
  */
 
 import Axios from 'axios';
-import { getCurrentLocaleCookie } from '@/lib/locale-cookie';
+import LocaleCookie from '@/lib/LocaleCookie';
 
 const api = Axios.create({
   baseURL: process.env.NEXT_PUBLIC_API_DOMAIN,
   headers: {
     'Accept': 'application/json',
-    'Accept-Language': getCurrentLocaleCookie(),
+    'Accept-Language': LocaleCookie.getCurrentLocaleCookie(),
     'Content-Type': 'application/json'
   }
 });
