+++ frontend/lib/apollo-client.js
@@ -10,7 +10,7 @@
 import { ApolloClient, ApolloProvider, createHttpLink, InMemoryCache } from '@apollo/client';
 import { setContext } from '@apollo/client/link/context';
 import { getAuthToken } from 'lib/auth';
-import { getCurrentLocaleCookie } from '@/lib/locale-cookie';
+import LocaleCookie from '@/lib/LocaleCookie';
 
 const httpLink = createHttpLink({
   uri: `${process.env.NEXT_PUBLIC_API_DOMAIN}/graphql`
@@ -23,7 +23,7 @@ const authLink = setContext(async (_, { headers }) => {
     headers: {
       ...currentHeaders,
       'authorization': auth ? auth.authorizationString : null,
-      'Accept-Language': getCurrentLocaleCookie()
+      'Accept-Language': LocaleCookie.getCurrentLocaleCookie()
     }
   };
 });
